aws_deployment/requirements.txt
-i https://pypi.org/simple
attrs==24.2.0; python_version >= '3.7'
aws-cdk-lib==2.18.0; python_version >= '3.6'
cattrs==24.1.2; python_version >= '3.8'
constructs==10.4.2; python_version ~= '3.8'
importlib-resources==6.4.5; python_version >= '3.8'
jsii==1.104.0; python_version ~= '3.8'
publication==0.0.3
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
typeguard==2.13.3; python_full_version >= '3.5.3'
typing-extensions==4.12.2; python_version >= '3.8'
